Rearrange mocks for network manager.
